Moved methods to MsgPatternHelper
